Test and implement AlphaBetaGammaFilter * Implement AlphaBetaGammaFilterTest * Implement AlphaBetaGammaFilter * Change EstimationFilter.calculatePrediction() to calculateStatePrediction() * Finish toString() of SystemState * Modify/add new constructor overloads to SystemState * Implement KalmanFilterEquations.useAccelerationStateUpdateEquation() * Implement KalmanFilterEquations.useVelocityStateExtrapolationEquation() * Test and implement overload to usePositionalStateExtrapolationEquation() for acceleration

+++ src/main/java/com/varunsingh/kalmanfilter/MeasureUpdatePredictFilter.java
@@ -5,7 +5,7 @@ abstract class MeasureUpdatePredictFilter implements EstimationFilter {
     protected double currentState;
     protected int iteration = 0;
 
-    MeasureUpdatePredictFilter(int initialStateGuess) {
+    MeasureUpdatePredictFilter(double initialStateGuess) {
         currentState = initialStateGuess;
     }
